refactor: move combine_audio_video method to ContentProcessor class
