+++ downloader.py
@@ -11,6 +11,7 @@ import threading
 
 # Local imports
 from content import Video, Audio
+from processor import ContentProcessor
 
 
 class YouTubeDownloader:
@@ -23,6 +24,9 @@ class YouTubeDownloader:
         self.video_path = os.path.join(self.base_path, "video")
         self.audio_path = os.path.join(self.base_path, "audio")
         self.combined_path = os.path.join(self.base_path, "combined")
+        self.processor = ContentProcessor(
+            self.video_path, self.audio_path, self.combined_path
+        )
         self._create_directories()
 
     def _create_directories(self):
@@ -140,20 +144,8 @@ class YouTubeDownloader:
             video_filename = f"{self.title}_video.mp4"
             audio_filename = f"{self.title}_audio.mp3"
             output_filename = f"{self.title}_combined.mp4"
-            self.combine_audio_video(video_filename, audio_filename, output_filename)
+            self.processor.combine_audio_video(
+                video_filename, audio_filename, output_filename
+            )
         else:
             print("Invalid choice. Exiting.")
-
-    def combine_audio_video(
-        self, video_filename: str, audio_filename: str, output_filename: str
-    ):
-        """Combines audio and video files into a single file"""
-        video_path = os.path.join(self.video_path, video_filename)
-        audio_path = os.path.join(self.audio_path, audio_filename)
-        output_path = os.path.join(self.combined_path, output_filename)
-
-        input_video = ffmpeg.input(video_path)
-        input_audio = ffmpeg.input(audio_path)
-
-        ffmpeg.concat(input_video, input_audio, v=1, a=1).output(output_path).run()
-        print(f"Combined video saved to: {output_path}")
